rename validate -> validation, add jsdoc
Committed-By-Agent: claude
