+++ .changeset/validate-programmatic-api.md
@@ -2,4 +2,4 @@
 'mppx': minor
 ---
 
-Added programmatic validation API (`mppx/validate` export), OpenAPI path parameter substitution, x402 protocol detection, multi-path discovery fallback, and `--outputJson` CLI flag.
+Added programmatic validation API (`mppx/validation` export), OpenAPI path parameter substitution, x402 protocol detection, multi-path discovery fallback, and `--outputJson` CLI flag.
