+++ vite.config.ts
@@ -21,7 +21,7 @@ const alias = {
   'mppx/stripe': path.resolve(import.meta.dirname, 'src/stripe'),
   'mppx/stripe/client': path.resolve(import.meta.dirname, 'src/stripe/client'),
   'mppx/stripe/server': path.resolve(import.meta.dirname, 'src/stripe/server'),
-  'mppx/validate': path.resolve(import.meta.dirname, 'src/validate'),
+  'mppx/validation': path.resolve(import.meta.dirname, 'src/validation'),
   mppx: path.resolve(import.meta.dirname, 'src'),
   '~test': path.resolve(import.meta.dirname, 'test'),
 }
